% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Example leaflet widget}
\usage{
read_file(filename)
}
\arguments{
\item{title}{label for the marker}

\item{lat}{lattitude coordinate}

\item{lng}{longintude coordinate}
}
\description{
Hello world of leaflet htmlwidget wrapped in OpenCPU. Based on example:
\url{http://rstudio.github.io/leaflet/}.
}
